♻️ refactor: separate displayClothes() to other file
